Provide consistent view of job metadata, expire after 14 days
First, resolve #81515 with JobMetadata to provide a single class where
job metadata is kept. JobMetadata has one interface but will use
redisson's live object service (on redis) when -Dwres.redisHost is
specified or fall back on caffeine soft-reference-eviction cache when
-Dwres.redisHost is not specified.

Second, keep a Map of job ids to JobMetadata. Again, single interface,
uses redisson's distributed collection (on redis) when
-Dwres.redisHost is specified or caffeine soft-reference-eviction
cache when -Dwres.redisHost is not specified.

Use TTL expiration on the Map<String,JobMetadata> when using redisson
to clear old jobs and an event handler to (attempt) to remove the job
metadata when expired from the Map.

Redis is therefore not required, but opens up new possibilities such
as running multiple taskers simultaneously sharing objects via redis.
This has not been implemented or tested, but the groundwork is here.

With persistence enabled on the redis instance, job metadata survives
tasker restarts.

Refs #55405
Fixes #81515