Distinguish filenames: one feature may be paired twice in one run
When there are only 1:1:1 feature correlations within a single
evaluation, WRES has the option to write only the left, only the
right, or only the baseline feature name and have all output names
avoid collision. But in order to support 1:N:1 or N:M:O feature
correlations in a single evaluation, WRES must disambiguate.

This commit does not address the contents of the output files, only
the file names.

Refs #81017
Refs #83234