Afford alternative netCDF statistics format 'netcdf2'
The netCDF blobs are built from scratch. When geospatial data using
coordinate reference system EPSG:4326 is found in a dataset, the
resulting netCDF output loads out-of-the-box with GDAL/OGR and QGIS.

As of this commit, the original template-based 'netcdf' statistics
format is still available unchanged to give users an opportunity to
test and migrate to the newer format without breaking changes.

As of this commit, the netcdf2 format cannot be used simultaneously
with the netcdf format because they both write to the same names. An
exception should occur if this is attempted.

Refs #83373