Allow a minimal version of DataSource within IngestResult
When no retry is needed, minimal source detail is needed to be sent
back to finish ingest (i.e. make wres.ProjectSource and wres.Project
rows). URIs were observed to take much of the heap, so the idea is
to let them be garbage collected so that millions of timeseries can
be ingested successfully in one job with under 2GiB heap.

Remove unused code, remove no-longer-used temp file references.

Refs #86147