Reading/ingest classes return CompletableFuture rather than Future
The short-range goal is to be able to use ProcessorHelper's
doAllOrException method or something like it at the top-level of
reading and ingest. Making more idiomatic use of CompletableFuture,
CompletionStage, Stream, etc. for reading and ingest is a
longer-range goal. Splitting ingest from readers entirely is also a
longer-range goal. The purpose of using doAllOrException is to have
Exceptions and Errors propagate quickly when multiple long-running
readers/ingesters are active.

Refs #86147 (found OOME not propagating quickly)
Refs #66545 (general Exception/Error-not-propagating-quickly ticket)