Request full forecast date range from WRDS AHPS services
The WRDS AHPS data will be de-duplicated per timeseries, so there is
no benefit to chunking by a particular date-range unless the WRDS
services were to do performance optimization around caching and the
cache headers (we are far from that).

For clarity and to more easily track progress, sort the date ranges.

Refs #80554
Refs #86887