Relax assertions of system test scenario 703
The goal of the test is to verify that data for hundreds of sites can
be successfully retrieved from NWIS and processed into statistics.
It is not to check the correctness of the data inside nor the
particular sites that were retrieved, only that hundreds were
retrieved. The reason for the relaxation is NWIS changes which sites
are available more frequently than one would expect and updating the
700 tests every time there is an NWIS change (or any WRES change) is
more cumbersome than one would expect.

Refs #87148