Make the dockerize script's curl commands directly copy/pastable
Make a little clearer that the script was intended to be run from a
build server with docker installed and has just finished building the
artifacts that should already be present. Given that we do not have
such a build server today, we run it manually as a stop-gap. But it
had a foot in each door already. We still can aim for automated runs.

Refs #88672