Change version of zips when an in-WRES module changes
This also changes the jar version when no code changes in that jar
actually happened, but this is still safer than the alternative.

There is probably a more gradle-y way to do this by accessing each
application's modules/components/dependencies as well.

Another improvement might be to split the application/zip blocks
away from the library/jar blocks, such that zip versions vary when
jar versions vary and jar versions only vary when code in the jar
varies.

Refs #88672