Make specific directories for specific roles and purposes
Give the tasker a place to write posted data, only for tasker to write
and for worker to read.

Give each process type a specific directory for heap dumps. This way
they won't accidentally overwrite each other's files (docker can give
the same pid because they are in different containers).

Refs #64736