Detect content type of source data automatically
When WRES is given simple file, directory, or InputStream data, it
detects whether it will be able to read the data source. There is no
more format attribute on the source declaration nor filename-based
switching when choosing a reader. The first few bytes of data are used
to either process further (e.g. gzip, tar data) or choose a reader
(e.g. datacard, pixml, wrds json, etc.) In the case where a single
timeseries is composed from multiple bytestreams (nwm) this is called
complex. In the case where a single source results in multiple
bytestreams each of which in turn is a simple source (where full
timeseries come from a single bytestream), at first it is called
complex but after decomposition each will be called by the name that
indicates a reader can process it. In other words, the disposition of
the data source changes as WRES processes the declared source down
through ingest.

* ReaderFactory now uses internal-to-WRES DataDisposition
* Remove declared source format from schema and all code
* Detection uses first 1024 bytes because < fs block and net frame
* Additional tarball detection when tika detects text/plain
* Tolerate a format declaration and log warnings for compatibility
* Declaration validation events now a member variable
* Separate tar handling from gzip handling in ZippedSource
* Move gzip handling outside of XMLReader class
* Clean up temp files created by ZippedSource
* CSVDataProvider uses UTF-8
* Close tika resources in tasker
* Update tests to reflect changes

Refs #38111 (Achieves "at best" resolution of the issue)
Refs #48054 (Resolves the issue)
Refs #64736 (Enables this and keeps tasker out of data inspection)