Add terminal job state FAILED_BEFORE_IN_QUEUE, rename others
It is anticipated that someday the declaration could be submitted
while the already-created post connections are in use (streaming).
That's why they were renamed in an effort to capture that data could
possibly still be sent in already-initiated posts, but that no new
posts would be permitted. Regardless, for this commit, first all the
timeseries need to be posted and then the postInputDone=true sent,
which sends the declaration to the job queue.

Refs #64736