Increase the main pipeline work queues, set connection timeout
If the queues cannot fit all the tasks created by the two loops in
FeatureProcessor and ProcessorHelper classes, then the rejected
execution handler causes the caller-run policy to be activated,
causing the main thread to take a long time to reach the method
Pipelines.doAllOrException(). Pipelines.doAllOrException() is what
allows one exception in one task to propagate quickly.

Separately, when logExecution() is called and the database is down,
instead of waiting forever, wait a long time and timeout. This can
be configured via wresconfig.xml database setting connectionTimeoutMs
or overridden by System Property wres.databaseConnectionTimeoutMs.

Refs #92650