Return HTTP 400 Bad Request when posted declaration is bad
This is when postInputDone=true is posted to the /{jobId}/input url.

Use ProjectConfigException for this purpose. When inputs is missing or
left or right is missing within inputs throw ProjectConfigException.
For IOException and other exceptions continue to return HTTP 500.

In any/all cases of failure, remove the already-posted input data.

Refs #94514