Turns on jfr flight recordings for the graphics clients in cowres/cluster mode.
There is a continuous recording for each client, which is contained within a subdirectory of /mnt/wres_share/heap_dumps/graphics. The subdirectory is named with the identifier of the container that wraps the graphics client. This directory is known as the jfr repository and contains recorded chunks of ~14 MiB each. To assemble a final recording from the chunks, use jfr assemble <repository> <filename.jfr>. There is also a dump-on-exit, which creates a final jfr recording when the jvm exits.

Refs #94065