Renders the variable declaration optional, except when reading from web services and NWM data sources (local or via web services).
When the variable is required and omitted, an error occurs in Validation.java (at the earliest) or when constructing a reader (at the latest), always before reading and ingest proper.

In general, our readers are not consistent with how they treat undeclared variables. When a variable is declared and the reader supplies that variable plus other variables, the TimeSeriesIngester does not drop the other variables, but ingests them. Dropping the other variables is a potential optimization in the TimeSeriesIngester when a variable is declared, but not implemented here.

Refs #95012