Use the Unified Code for Units of Measure (UCUM)
See https://ucum.org/ucum.html

* UCUM aims to be comprehensive for machine-to-machine data exchange.
* UCUM is a mature standard.
* UCUM uses plain 7-bit ASCII for unit representations.
* See a handy tool at https://ucum.nlm.nih.gov/ucum-lhc/demo.html
* WRES can reference UCUM and existing demos/tools rather than indriya

The use of UCUM applies to the values (not keys) in our convenience
aliases map and also to the units (not alias) in <unitAlias>
declarations. The <unitAlias> declarations disambiguate and clarify
exactly which unit is intended for a given vernacular/alias string.

For example, @miles/hour@ may be declared to mean UCUM @[mi_i]/h@ by
use of the <unitAlias> tag.

See User Support project documentation as well as the code associated
with this commit as well as the ticket associated with this commit.

Refs #44819