SQLite supports "limit" clauses
But WRES does not yet support SQLite, MariaDB, or MySQL. Most of the
liquibase database migration changesets do not prohibit any of these
databases and should work but have not been tested. Changes in and
around wres.system and wres.io.utilities classes would be required to
experiment with support for additional database systems. The WRES
project has prioritized support for a central WRES cluster, on which
we run postgres. It is possible that folks who wish to run WRES on
their own machines in future would already have their own DB systems
that they prefer, in which case they or the WRES team could make the
handful of changes to experiment with those database systems.

Support for anything other than PostgreSQL or H2 is not a priority at
this time, but this little bit of knowledge (SQLite supports limit
clauses) does not need to be lost.

Refs #103431