Prevent jaxb-core from being on the runtime classpath
This avoids conflicting classes provided by jaxb-core and jaxb-runtime
being on the same classpath. The wres-config project has the jaxb 2
stuff because immutable-xjc is stuck at jaxb 2 for now. The
wres-tasker project had some jaxb 3 library declared but unnecessarily
because it was only used in one place. After replacing the use in
JobStandardStreamConsumer with a commons-codec method, the tasker code
is happy with the reduction in scope/visibility of jaxb-core to the
compileOnlyApi configuration.

Refs #103431 (found and fixed during the course of this ticket)