Group metrics by their common parameters.
Allow for the grouping of metrics in the new declaration language
by their common parameters. Each group is an atomic group for
processing. Unit test the grouping. Re-scope the minimum_sample_size
to the top-level of the declaration, which is the 99% use case. If
a per-metric parameter is needed in future, allow it.

Refs #113677