Fixing Database Lock
refs #121563

We were locking the database before we were swapping it in the clean
method. The lead to us locking the active database while trying to clean
the non-active one and was leading to the clean/switch to fail along
with jobs to fail due to deadlock situations.

Also added some code to log the response body of failed evaluations