Writing more output after job is closed
refs # 121563

Output was sometimes cut off too soon. This is change is the best we can
do. Since the output is tied to an evaluation status we only have 10
seconds to write the rest of the output. This should make things much
better, but until we can tie the output of an evaluation to an object
that has an end this is the best we can do