Defaulting content type to null
refs #121850

When an evaluation ends and someone is retieving files programatically
we were setting a return type before sending it over. In the case that
we were not able to determine a type we were setting it to TEXT as a
default. This was causing csvt files to not be able to be fetched.
Defaulting to no type being set