Optimizing log writing on worker server
refs #127827

Every time we were in a while loop we were reading bytes from a buffer
into a new string object and checking if that string was empty before
writing to our output buffer. We now only create a new string when there
are new bytes in the buffer. This should save on heap space and
unnecissary string intantiation calls